third_party/blink/web_tests/platform/linux/external/wpt/paint-timing/fcp-only/fcp-broken-image-expected.txt

This is a testharness.js-based test.
[FAIL] Broken image should not trigger First Contentful Paint.
  assert_equals: First contentful paint marked too early.  expected 0 but got 1
Harness: the test ran to completion.
